refactor - #122 ChallengeAppService 추가
